[meta-f] fix passing the same argument twice
When placing function arguments in registers, cps_to_bytecode just
swaps/moves the arguments in place forgetting their previous
location. If argument was already passed in one of the previous
registers, a garbage can be swapped in its place.

Make compiler handle detect registers that are passed multiple times
and stash variables in the way to some different location.